I have made an extensive User class
i have made the user class so that it is able to send and recieve the
messages itself. it is able to send and receive all the messages (not
functional yet though). the serverapp is also able to listen for a
connection and add a user to the chat -- i will fine tune those details
later to allow for multiple users.